pyproject.toml
[project]
name = "metraq_dip"
version = "0.1.0"
description = "METRAQ spatial inference project"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.1",
    "kaleido>=1.2.0",
    "numpi>=0.9.0",
    "pandas>=2.3.3",
    "pandas-stubs>=2.2.2.240807",
    "plotly>=6.5.2",
    "pykrige>=1.7.3",
    "pymysql>=1.1.2",
    "pyproj>=3.6.1",
    "pyyaml>=6.0.3",
    "scipy-stubs~=1.15.3",
    "shapely>=2.0.7",
    "sqlalchemy>=2.0.46",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-cov",
  "pytest-xdist",
  "setuptools>=80.10.2",
]
docs = [
  "mkdocs>=1.6.1",
]

[project.scripts]
train = "pipelines.main:main"

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
